LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/worker-set-timeout.sub-expected.txt


FAIL Dedicated worker inherits CSP from main document assert_equals: expected "setTimeout allowed" but got "setTimeout blocked"
FAIL Shared worker delivers its own CSP Can't find variable: SharedWorker
